Remove base viewModel for toolSettings Remove viewModels for toolSettings choose language and move logic into ToolSettingsViewModel
