--- godtools/App/Features/ToolSettings/Presentation/ToolSettings/BaseToolSettingsViewModel.swift
@@ -1,32 +1,0 @@
-//
-//  BaseToolSettingsViewModel.swift
-//  godtools
-//
-//  Created by Levi Eggert on 5/12/22.
-//  Copyright © 2022 Cru. All rights reserved.
-//
-
-import Foundation
-
-class BaseToolSettingsViewModel: ObservableObject {
-    
-    @Published var hidesShareables: Bool = false
-    
-    init() {}
-    
-    func getTopBarViewModel() -> BaseToolSettingsTopBarViewModel {
-        return BaseToolSettingsTopBarViewModel()
-    }
-    
-    func getOptionsViewModel() -> BaseToolSettingsOptionsViewModel {
-        return BaseToolSettingsOptionsViewModel()
-    }
-    
-    func getChooseLanguageViewModel() -> BaseToolSettingsChooseLanguageViewModel {
-        return BaseToolSettingsChooseLanguageViewModel()
-    }
-    
-    func getShareablesViewModel() -> BaseToolSettingsShareablesViewModel {
-        return BaseToolSettingsShareablesViewModel()
-    }
-}
