+++ godtools/App/Features/ToolSettings/Presentation/ToolSettings/ToolSettingsViewModel.swift
@@ -9,7 +9,7 @@ import Foundation
 import GodToolsToolParser
 import Combine
 
-class ToolSettingsViewModel: BaseToolSettingsViewModel {
+class ToolSettingsViewModel: ObservableObject {
     
     private let manifestResourcesCache: ManifestResourcesCache
     private let localizationServices: LocalizationServices
@@ -18,7 +18,15 @@ class ToolSettingsViewModel: BaseToolSettingsViewModel {
     private let trainingTipsEnabled: Bool
     private let shareables: [Shareable]
     
+    private var primaryLanguageCancellable: AnyCancellable?
+    private var parallelLanguageCancellable: AnyCancellable?
+    
     private weak var flowDelegate: FlowDelegate?
+    
+    @Published var chooseLanguageTitle: String = ""
+    @Published var primaryLanguageTitle: String = ""
+    @Published var parallelLanguageTitle: String = ""
+    @Published var hidesShareables: Bool = false
         
     required init(flowDelegate: FlowDelegate, manifestResourcesCache: ManifestResourcesCache, localizationServices: LocalizationServices, primaryLanguageSubject: CurrentValueSubject<LanguageModel, Never>, parallelLanguageSubject: CurrentValueSubject<LanguageModel?, Never>, trainingTipsEnabled: Bool, shareables: [Shareable]) {
         
@@ -29,13 +37,34 @@ class ToolSettingsViewModel: BaseToolSettingsViewModel {
         self.parallelLanguageSubject = parallelLanguageSubject
         self.trainingTipsEnabled = trainingTipsEnabled
         self.shareables = shareables
-        
-        super.init()
-        
+        chooseLanguageTitle = localizationServices.stringForMainBundle(key: "toolSettings.chooseLanguage.title")
         hidesShareables = shareables.isEmpty
+        
+        primaryLanguageCancellable = primaryLanguageSubject.sink(receiveValue: { [weak self] (language: LanguageModel) in
+            
+            guard let weakSelf = self else {
+                return
+            }
+            
+            weakSelf.primaryLanguageTitle = weakSelf.getTranslatedLanguageName(language: language)
+        })
+        
+        parallelLanguageCancellable = parallelLanguageSubject.sink(receiveValue: { [weak self] (language: LanguageModel?) in
+            
+            guard let weakSelf = self else {
+                return
+            }
+            
+            if let language = language {
+                weakSelf.parallelLanguageTitle = weakSelf.getTranslatedLanguageName(language: language)
+            }
+            else {
+                weakSelf.parallelLanguageTitle = weakSelf.localizationServices.stringForMainBundle(key: "toolSettings.chooseLanguage.noParallelLanguageTitle")
+            }
+        })
     }
     
-    override func getTopBarViewModel() -> BaseToolSettingsTopBarViewModel {
+    func getTopBarViewModel() -> BaseToolSettingsTopBarViewModel {
         
         guard let flowDelegate = flowDelegate else {
             assertionFailure("Failed to instantiate viewModel, flowDelegate should not be nil.")
@@ -48,7 +77,7 @@ class ToolSettingsViewModel: BaseToolSettingsViewModel {
         )
     }
     
-    override func getOptionsViewModel() -> BaseToolSettingsOptionsViewModel {
+    func getOptionsViewModel() -> BaseToolSettingsOptionsViewModel {
         
         guard let flowDelegate = flowDelegate else {
             assertionFailure("Failed to instantiate viewModel, flowDelegate should not be nil.")
@@ -62,22 +91,7 @@ class ToolSettingsViewModel: BaseToolSettingsViewModel {
         )
     }
     
-    override func getChooseLanguageViewModel() -> BaseToolSettingsChooseLanguageViewModel {
-        
-        guard let flowDelegate = flowDelegate else {
-            assertionFailure("Failed to instantiate viewModel, flowDelegate should not be nil.")
-            return BaseToolSettingsChooseLanguageViewModel()
-        }
-        
-        return ToolSettingsChooseLanguageViewModel(
-            flowDelegate: flowDelegate,
-            localizationServices: localizationServices,
-            primaryLanguageSubject: primaryLanguageSubject,
-            parallelLanguageSubject: parallelLanguageSubject
-        )
-    }
-    
-    override func getShareablesViewModel() -> BaseToolSettingsShareablesViewModel {
+    func getShareablesViewModel() -> BaseToolSettingsShareablesViewModel {
         
         guard let flowDelegate = flowDelegate else {
             assertionFailure("Failed to instantiate viewModel, flowDelegate should not be nil.")
@@ -90,5 +104,24 @@ class ToolSettingsViewModel: BaseToolSettingsViewModel {
             manifestResourcesCache: manifestResourcesCache,
             localizationServices: localizationServices
         )
+    }
+    
+    private func getTranslatedLanguageName(language: LanguageModel) -> String {
+        return LanguageViewModel(language: language, localizationServices: localizationServices).translatedLanguageName
+    }
+    
+    func primaryLanguageTapped() {
+        
+        flowDelegate?.navigate(step: .primaryLanguageTappedFromToolSettings)
+    }
+    
+    func parallelLanguageTapped() {
+        
+        flowDelegate?.navigate(step: .parallelLanguageTappedFromToolSettings)
+    }
+    
+    func swapLanguageTapped() {
+        
+        flowDelegate?.navigate(step: .swapLanguagesTappedFromToolSettings)
     }
 }
