+++ godtools/App/Features/ToolSettings/Presentation/ToolSettings/ToolTraining/Transitioning/ToolTrainingAnimatedTransitioning.swift
