+++ godtools/App/Features/ToolSettings/Presentation/ToolSettings/ToolTraining/Views/ToolTrainingTipView.swift
